+++ packages/tokens/src/colors.ts
@@ -60,7 +60,6 @@ export const colors = {
   P400: '#ED55C2',
   P500: '#BE449B',
   P600: '#8E3374',
-  Barbie: '#8E3374',
   O50: '#FDF7F4',
   O100: '#F8E3DA',
   O200: '#F2C8B6',
@@ -68,5 +67,4 @@ export const colors = {
   O400: '#C34F16',
   O500: '#B25E3A',
   O600: '#85462B',
-  O700: '#85462C',
 }